🚚Rename: model폴더에 있는 타입과 스키마를 크게 분리 후 연관된 타입별로 분류
