+++ src/Components/auth/PersonalInfoForm.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { UseFormReturn } from "react-hook-form";
-import { TSignupFormData } from "@/model";
+import { TSignupFormData } from "@/model/types/authTypes/signup.type";
 import { FormField, FormItem, FormLabel, FormControl, FormMessage } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
