+++ src/Components/auth/PositionSelector.tsx
@@ -1,4 +1,4 @@
-import { TPosition } from "@/model";
+import { TPosition } from "@/model/types/position.type";
 import { Label } from "@/components/ui/label";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Briefcase, User } from "lucide-react";
