+++ src/hooks/useCompanySettingForm.ts
@@ -1,6 +1,6 @@
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { companyFormSchema } from "@/model/managerFirstSchema";
+import { companyFormSchema } from "@/model/schema/managerFirstSchema/managerFirst.schema";
 
 export const useCompanySettingForm = () => {
   return useForm({
