+++ src/hooks/useSignup.ts
@@ -6,7 +6,9 @@ import { useUserStore } from "@/store/user.store";
 import { signupFormSchema } from "@/model/schema/authSchema/signup.schema";
 import { getCompanyInfo, validateCompanyCode } from "@/api";
 import { signup } from "@/api/auth";
-import type { TSignupFormData, TPosition, TEmpUserData, TCMUserData } from "@/model";
+import { TSignupFormData } from "@/model/types/authTypes/signup.type";
+import { TPosition } from "@/model/types/position.type";
+import { TEmpUserData, TCMUserData } from "@/model/types/user.type";
 import { z } from "zod";
 import { useCompanyStore } from "@/store/company.store";
 import { nanoid } from "nanoid";
