+++ src/model/types/position.type.ts
@@ -1,0 +1,2 @@
+export type TPosition = "manager" | "employee";
+export type TEmploymentType = "정규직" | "계약직" | "일용직" | "선택안함";
